Drop Ansible Python API
While I'm in favor of using SDKs and APIs whenever
is possible, in this case I got the feeling that
it's not documented well enough for to use it.

I went with the approach many big projects took
(including Zuul itself) which is running it directly
through the supported CLI.

In addition this change takes case of cloning Zuul
which includes the modules required for reproducing
some of the jobs.