Copy project to host
Zuul jobs eventually use the project in order to test it
or run code (like some scripts) from it.

Without transferring the project to the host on which the job
operates we can expect the job to fail.

This change handles it by transferring it from local path
to the same path on the host.